+++ src/podofo/base/PdfFontMetrics.cpp
@@ -347,7 +347,7 @@ const FreeTypeFacePtr& PdfFontMetricsBase::GetFaceHandle() const
         auto& rthis = const_cast<PdfFontMetricsBase&>(*this);
         auto view = GetFontFileDataHandle().view();
         FT_Face face;
-        if (view.size() != 0 && PoDoFo::TryCreateFreeTypeFace(view, face))
+        if (view.size() != 0 && FT::TryCreateFaceFromBuffer(view, face))
             rthis.m_Face = FreeTypeFacePtr(face);
         else
             rthis.m_Face = FreeTypeFacePtr();