+++ src/podofo/base/PdfFontMetricsFreetype.h
@@ -25,7 +25,7 @@ class PODOFO_API PdfFontMetricsFreetype final : public PdfFontMetrics
 public:
     static std::unique_ptr<PdfFontMetricsFreetype> FromMetrics(const PdfFontMetrics& metrics);
 
-    static std::unique_ptr<PdfFontMetricsFreetype> FromBuffer(const charbuff::const_ptr& buffer);
+    static std::unique_ptr<PdfFontMetricsFreetype> FromBuffer(const std::shared_ptr<const charbuff>& buffer);
 
     /// <summary>
     /// Create a metrics from a FT_Face